requirements.txt
streamlit
sqlalchemy
fpdf
Pillow